# packages/backend/openapi/openapi.yaml
...
openapi: '3.1.0'
info:
  title: Linked Data Explorer Backend
  description: >-
    DMN discovery and chain execution, SHACL validation, asset storage and
    integrations for the Linked Data Explorer.

    Every 4xx/5xx response is `application/problem+json` (RFC 9457,
    schema `Problem`): `status`, `title` and `detail` always; `code`, the
    machine-readable identifier existing callers branch on, where the
    handler sets one. Two operations are deliberate exceptions, each noted
    at its own response: `POST /dmns/evaluate/{decisionKey}` is a byte-for-byte
    proxy for Operaton's own evaluate endpoint and passes Operaton's error
    body through unchanged (or answers `OperatonProxyError` when Operaton
    could not be reached at all), and `GET /health`'s 503 body is the health
    document itself, not a problem response, because the document *is* the
    explanation. https://github.com/sgort/linked-data-explorer/issues/131.
  contact:
    name: Steven Gort
    email: steven.gort@ictu.nl
    url: https://iou-architectuur.open-regels.nl/
servers:
  - url: https://backend.linkeddata.open-regels.nl/v1
    description: Production
  - url: https://acc.backend.linkeddata.open-regels.nl/v1
    description: Acceptance
...
paths:
...
  /chains:
    get:
      operationId: discoverChains
      tags: [Execution]
      summary: Pairwise DMN chain links, grouped by source DMN
      description: >-
        Discovers, via SPARQL, every DMN whose declared output variable
        matches another DMN's declared input variable, then groups those
        pairwise links by the source (`from`) DMN. Takes no parameters —
        `req.query` is never read — so there is nothing for the handler to
        reject with a 400. `chains` is `[]`, still with a 200, when no links
        are found.
      responses:
        '200':
          description: The chain links, grouped by source DMN. Empty when none exist.
          headers:
            API-Version:
              $ref: '#/components/headers/ApiVersion'
          content:
            application/json:
              schema:
                type: object
                required: [success, data, timestamp]
                properties:
                  success:
                    const: true
                  data:
                    type: object
                    required: [total, chains]
                    properties:
                      total:
                        type: integer
                        minimum: 0
                        description: '`chains.length`.'
                      chains:
                        type: array
                        items:
                          $ref: '#/components/schemas/ChainMapEntry'
                  timestamp:
                    type: string
                    format: date-time
        '500':
          description: The SPARQL lookup failed (code DISCOVERY_ERROR).
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  /chains/execute:
    post:
      operationId: executeChain
      tags: [Execution]
      summary: Execute a chain of DMNs sequentially, or a pre-assembled DRD in one call
      description: >-
        Threads each DMN's outputs into the next DMN's inputs, evaluating each
        one in Operaton in turn (`orchestrationService.executeChain`). When
        `isDrd` and `drdEntryPointId` are both given, that sequential
        execution is skipped entirely and all `inputs` are sent to Operaton's
        evaluation of the pre-assembled DRD in a single call instead. A step
        is only ever added to `steps` once it has completed successfully —
        `orchestration.service.ts` pushes it after its try/catch, and a step
        whose evaluation throws re-throws before that push is reached, so it
        is never added; the problem's `detail` carries that failure instead.
        Every populated `steps` entry therefore always has `outputs`, never
        `error`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [dmnIds, inputs]
              properties:
                dmnIds:
                  type: array
                  minItems: 1
                  items:
                    type: string
                  description: >-
                    DMN identifiers to execute in order (or, for the DRD path,
                    the identifiers the DRD was assembled from). Required and
                    must not be empty.
                inputs:
                  type: object
                  minProperties: 1
                  description: >-
                    Free-form initial input variables. Required and must not
                    be an empty object.
                options:
                  type: object
                  description: >-
                    Only `includeIntermediateSteps` is read; `timeout` is
                    declared in the request type but never read anywhere and
                    has no effect.
                  properties:
                    includeIntermediateSteps:
                      type: boolean
                      description: When true, the response's `data` carries `steps`.
                endpoint:
                  type: string
                  description: >-
                    SPARQL endpoint URL to query instead of the configured
                    one, used to look up the chain's DMNs. Unvalidated, and
                    ignored when `isDrd` is true, because that path does not
                    look the DMNs up.
                isDrd:
                  type: boolean
                  description: >-
                    When true together with `drdEntryPointId`, executes the
                    pre-assembled DRD instead of the sequential chain.
                drdEntryPointId:
                  type: string
                  description: >-
                    The DRD's entry-point decision key, used only when `isDrd`
                    is also true.
      responses:
        '200':
          description: The chain (or DRD) completed successfully.
          headers:
            API-Version:
              $ref: '#/components/headers/ApiVersion'
          content:
            application/json:
              schema:
                type: object
                required: [success, data, timestamp]
                properties:
                  success:
                    const: true
                  data:
                    type: object
                    required: [success, chainId, executionTime, finalOutputs]
                    properties:
                      success:
                        const: true
                      chainId:
                        type: string
                        description: >-
                          The executed DMN identifiers joined with `->`, or,
                          for the DRD path, `DRD:<drdEntryPointId>`.
                      executionTime:
                        type: number
                        description: Milliseconds.
                      finalOutputs:
                        type: object
                        description: >-
                          Free-form: the accumulated chain variables once
                          every step has run, already unwrapped from
                          Operaton's `{value, type}` form.
                      steps:
                        type: array
                        description: Present only when the caller set `options.includeIntermediateSteps`.
                        items:
                          $ref: '#/components/schemas/ExecutionStep'
                  timestamp:
                    type: string
                    format: date-time
        '400':
          description: >-
            `dmnIds` is missing or empty (message "dmnIds array is required
            and must not be empty"), or `inputs` is missing or an empty
            object (message "inputs object is required"). Both are code
            INVALID_REQUEST.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '500':
          description: >-
            Three causes share this status. The orchestrator resolving with
            `success: false` (a DMN in the chain threw, or the DRD
            evaluation threw) answers a problem response whose `data`
            extension member carries the partial chain result -- which
            steps ran, the outputs so far -- so that information is not
            lost alongside the failure (`detail`). The route's own
            try/catch (code EXECUTION_ERROR) and the request body not being
            valid JSON, or over 10 MB (code INTERNAL_ERROR; #143 — the
            latter never reaches this route's own handler; it is answered
            by the global error handler) answer a problem response with no
            `data`.
          content:
            application/problem+json:
              schema:
                anyOf:
                  - allOf:
                      - $ref: '#/components/schemas/Problem'
                      - type: object
                        description: >-
                          The orchestrator resolved with `success: false`.
                          `data` carries the same fields the 200 response's
                          `data` does, minus `success`.
                        required: [data]
                        properties:
                          data:
                            type: object
                            required: [chainId, executionTime, finalOutputs]
                            properties:
                              chainId:
                                type: string
                              executionTime:
                                type: number
                                description: Milliseconds.
                              finalOutputs:
                                type: object
                                description: >-
                                  Free-form: whatever chain variables had
                                  accumulated before the failure. `{}` for a
                                  failed DRD run.
                              steps:
                                type: array
                                description: Present only when the caller set `options.includeIntermediateSteps`. `[]` for a failed DRD run.
                                items:
                                  $ref: '#/components/schemas/ExecutionStep'
                  - $ref: '#/components/schemas/Problem'
...
components:
  headers:
    ApiVersion:
      description: Full release version of the API (/core/version-header).
      required: true
      schema:
        type: string
        examples:
          - '2026.09.4'
  schemas:
    Problem:
      type: object
      description: >-
        RFC 9457 problem details (https://www.rfc-editor.org/rfc/rfc9457),
        returned as `application/problem+json` by every 4xx/5xx response
        except two: `POST /dmns/evaluate/{decisionKey}` forwards Operaton's
        own error body, and `GET /health`'s 503 is the health document
        itself -- both documented at their own operation. `code` is kept as
        an extension member alongside the RFC's own fields: the
        machine-readable identifier existing callers already branch on.
        https://github.com/sgort/linked-data-explorer/issues/131.
      required: [status, title, detail]
      properties:
        type:
          type: string
          format: uri-reference
          description: >-
            A URI identifying the problem type. `about:blank` (RFC 9457
            §4.2.1) unless an operation's description says otherwise --
            every problem is that operation's own generic 4xx/5xx, not a
            documented category with its own reference page.
          examples: [about:blank]
        status:
          type: integer
          description: The HTTP status, repeated from the response line.
        title:
          type: string
          description: >-
            A short, human-readable summary of the kind of problem. Stable
            for that kind: the same title recurs across requests and, where
            there is a `code`, across every response that sets that code.
          examples: [List failed]
        detail:
          type: string
          description: What went wrong with this request.
        instance:
          type: string
          description: The request path that produced this response.
          examples: [/v1/assets/bpmn]
        code:
          type: string
          description: >-
            The machine-readable identifier this operation's description
            names, kept for callers that already branch on it. Omitted by
            the few operations that never set one.
          examples: [LIST_FAILED]
...
    ChainMapEntry:
      type: object
      description: >-
        One source DMN and every downstream DMN it can feed, built by
        grouping the pairwise SPARQL chain links by their `from` DMN
        (chain.routes.ts, the handler's own local `ChainMapEntry`
        interface). Distinct from the DMN-record schemas of phase 2
        (DmnModel, DmnReference, EnhancedChainLink, ChainCycle).
      required: [from, connections]
      properties:
        from:
          type: string
          description: Source DMN identifier, from a SPARQL binding.
        connections:
          type: array
          items:
            type: object
            required: [to, variable, variableType]
            properties:
              to:
                type: string
                description: Downstream DMN identifier, from a SPARQL binding.
              variable:
                type: string
                description: The variable identifier that connects the two DMNs.
              variableType:
                type: string
                description: >-
                  The variable's declared type, from a SPARQL binding — a
                  plain string, never an enum.
    ExecutionStep:
      type: object
      description: >-
        One DMN evaluated within a chain execution. orchestration.service.ts
        only pushes a step onto `steps` after it has completed successfully;
        a step whose evaluation throws sets its own `error` field but is then
        discarded, re-thrown to the caller before that push runs, so it is
        never added to `steps`. Every step this API actually returns
        therefore always carries `outputs`, and `error` never appears here —
        unlike the `ExecutionStep` TypeScript type, which marks `outputs` and
        `error` as mutually-exclusive optionals.
      required: [dmnId, dmnTitle, startTime, endTime, duration, inputs, outputs]
      properties:
        dmnId:
          type: string
        dmnTitle:
          type: string
        startTime:
          type: number
          description: Epoch milliseconds.
        endTime:
          type: number
          description: Epoch milliseconds.
        duration:
          type: number
          description: Milliseconds; `endTime` minus `startTime`.
        inputs:
          type: object
          description: >-
            Free-form snapshot of the accumulated chain variables immediately
            before this step ran.
        outputs:
          type: object
          description: >-
            Free-form: this step's outputs, already unwrapped from Operaton's
            `{value, type}` form.
...